Flesh out serverside simulation
The biggest change in this commit is the mutation of a Command from
just an input the player sends us to a way of communicating with the
simulation in other ways as well, such as adding or removing players
from it.  These commands are packaged into an array and are used to
tick a snapshot into new state, and then sent to the clients so they
can both validate their past inputs and predict future inputs of other
players.