Server keeps track of input buffer as "health"
This is much more accurate than a mere round-trip-time, as there is no
better way to ensure we always have an input to the server in time than
by having the server tell the client how far ahead or behind it is and
the client moves itself in response.

I probably need something like a PID controller to regulate the buffer.