Replace wall ejection method with a better one
The correct way to eject a circle from a hit wall is to create a normal
vector going from the touch location to the center of the entity.  This
gives you the exact same ejection angle as before on flat walls, but on
corners you never have to deal with ambiguity of which wall you're
ejecting yourself from because you're just dealing with the collision
point in isolation.

Some of the tests no longer pass, but that's because the new method
doesn't quite do the same thing as before, so I need to rewrite those
tests to reflect the new behavior.