Change license to zlib
I always intended to make the engine zlib, and after some thought, I
realized that I was so far away from an actual game that I might as well
make the license switch now.

I also changed the slogan to something that is clearer about what the
engine is and does.